requirements.txt
datasets==4.3.0
gdown==5.2.0
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
Pillow==11.3.0
scikit_learn==1.5.1
timm==1.0.21
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.6
xgboost==2.1.1
